+++ swift-currency/Sources/ExchangeRate/ExchangeRateModel.swift
@@ -3,7 +3,7 @@ import Foundation
 struct ExchangeRate: Codable {
     let from: String
     let to: String
-    let rate: Decimal
+    let rate: Double
     let updatedAt: Date
 
     enum CodingKeys: String, CodingKey {
